XMP-43: cleanup;
